sunxi: add smc service handler for an SMC mailbox
The SCPI protocol uses a combination of a shared memory region and a
mailbox to signal requests from the application cores to the management
processor. However the Allwinner mailbox device in the A64 SoC can only
signal requests between the ARM cores and the OpenRISC core, not between
the ARM cores themselves.
As an alternative we can directly trigger SCPI requests by using an
smc instruction to call routines in the runtime code of ARM Trusted
Firmware (similar to what PSCI does).
Add a vendor defined (SiP) runtime service handler, which dispatches
incoming smc calls to the SCPI handler if the function ID matches.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>