sunxi: SCPI: implement MMC clocks
The interface clock speed at which an MMC controller communicates
with the storage controller is variable and depends on the capability
of both the host controller and the storage device.
Add the clocks for the three MMC controller to the SCPI clock
implementation to allow a non-secure OS to program the clock frequency
without knowing about the sunxi clock specifics.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>