sunxi: power: refactor power specific code into extra file
At the moment the power code mixes the code to access the RSB bus with
the actual PMIC register control and setup via this bus.
Separate those two to make the code more readable, also to provide
better abstraction.
This for instance later allows using I2C to control different PMICs,
for instance.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>