sunxi: clocks: implement CPU clock and export as an SCPI clock
Add functions to read and set the frequency of the (single) CPU clock,
which is done via a register controlling the CPUX PLL.
This replaces the hardcoded reset value for the safe 816 MHz by a call
to that function.
Also export the CPU clock via the SCPI interface, so that any OS can
easily change that clock.

Signed-off-by: Andre Przywara <andre.przywara@arm.com>